clean up unused code in ceph_medic.remote
