+++ ceph_medic/remote/__init__.py
@@ -1,5 +1,2 @@
-from . import mon  # noqa
-from . import osd  # noqa
-from . import common  # noqa
 from . import functions  # noqa
 from . import commands  # noqa
